+++ metro-ai-suite/metro-sdk-manager/scripts/oep-gen-ai-sdk.sh
@@ -23,14 +23,13 @@ repositories=(
 )
 
 images=(
-  "intel/audio-analyzer:2026.2.0-rc1"
-  "intel/document-ingestion:2026.1.0"
+  "intel/audio-analyzer:2026.2.0-rc2"
+  "intel/document-ingestion:2026.2.0-rc1"
   "intel/multimodal-embedding-serving:2026.2.0-rc1"
-  "intel/vdms-dataprep:2026.1.0"
-  "intel/vlm-openvino-serving:2026.2.0-rc1"
+  "intel/multimodal-dataprep:2026.2.0-rc1"
   "intel/model-download:2026.2.0-rc1"
-  "intel/chatqna:2026.1.0"
-  "intel/chatqna-ui:2026.1.0"
+  "intel/chatqna:2026.2.0-rc1"
+  "intel/chatqna-ui:2026.2.0-rc1"
 )
 NAME="OEP Gen AI SDK"